refactor(file-explorer): dedupe file icon/color mapping into utils
file-explorer.tsx defined getFileVisual()/FileIcon/DEFAULT_FILE_COLOR
locally, and file-tree-row.tsx redeclared the identical FileIcon type
just to type its fileVisual prop. Moved the mapper next to the sibling
getLanguageFromPath() in utils.ts (same extension-classification
pattern), and both consumers now import the single definition.
